Show Laravels validation messages
By using `getValidationRulesOnStore` the error messages don't contain
the field name. An example rule: `'name' => 'required,min:10'`. The
validation message will because `required,min:10`. But that is not
helpfull for the users. By returning an empty array the messages won't
be overwritten and "Field name is required" will be shown.